% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc386function.R
\name{Calc386}
\alias{Calc386}
\title{Calculate the U238/Pb206 ratio}
\usage{
Calc386(t, ..., L8 = 1.55125 * 10^-10)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the U238/Pb206 ratio for a given time, either by using
Stacey & Kramers (1975) 2nd stage model parameters as default, or by setting
optional arguments.
}
