% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc64function.R
\name{Calc64}
\alias{Calc64}
\title{Calculate the 206Pb/204Pb ratio}
\usage{
# Basic usage accepting default model parameters for Stacey & Kramers (1975) 2nd stage model.
  Calc64(t, ...)

# Advanced usage with optional arguments
  Calc64(t, T1, X1, Mu1, L8)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Mu1}{238U/204Pb model. Default is 9.74}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the 206Pb/204Pb ratio for a given time, either by using
Stacey & Kramers (1975) 2nd stage model parameters as default, or by setting
optional arguments.
}
\examples{
# Example of basic method using a hypothetical sample age of 2700 Ma
  Calc64(2700)

# Example of using optional arguments based on Stacey & Kramers (1975) 1st stage model
  Calc64(2700, T1 = 4570, X1 = 9.307, Mu1 = 7.19)

# Example of using the optional argument for Mu1 = 8 only, and accepting the other defaults
  Calc64(2700, Mu1 = 8)
}
