% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc84function.R
\name{Calc84}
\alias{Calc84}
\title{Calculate the 208Pb/204Pb ratio}
\usage{
# Basic usage accepting default model parameters for Stacey & Kramers (1975) 2nd stage model
  Calc84(t, ...)

# Advanced usage with optional arguments
  Calc84(t, T1, Z1, W1, L2)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{Z1}{208Pb/204Pb model starting composition. Default is 31.23}

\item{W1}{232Th/204Pb model. Default is 36.84}

\item{L2}{232Th decay constant. Default is 0.49475 * 10^-10}
}
\description{
This function calculates the 208Pb/204Pb ratio for a given time, either by using
Stacey & Kramers (1975) 2nd stage model parameters as default, or by setting
optional arguments.
}
\examples{
# Example of basic method using a hypothetical sample age of 2700 Ma
  Calc84(2700)

# Example of using optional arguments based on Stacey & Kramers (1975) 1st stage model
  Calc84(2700, T1 = 4570, Z1 = 29.487, W1 = 32.21)

# Example of using the optional argument for W1 = 30 only, and accepting the other defaults
  Calc84(2700, W1 = 30)
}
