% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc84infunction.R
\name{Calc84in}
\alias{Calc84in}
\title{Calculate the initial 208Pb/204Pb ratio}
\usage{
# Basic usage accepting default model parameters for Stacey & Kramers (1975) 2nd stage model
  Calc84in(t, x, y, z, ...)

# Advanced usage with optional arguments
  Calc84in(t, x, y, z, T1, X1, Y1, Z1, U8U5, L5, L8, L2)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{x}{sample 206Pb/204Pb ratio}

\item{y}{sample 207Pb/204Pb ratio}

\item{z}{sample 208Pb/204Pb ratio}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Y1}{207Pb/204Pb model starting composition. Default is 12.998}

\item{Z1}{208Pb/204Pb model starting composition. Default is 31.23}

\item{U8U5}{Present-day 238U/235U ratio. Default is 137.88}

\item{L2}{232Th decay constant. Default is 0.49475 * 10^-10}

\item{L5}{235U decay constant. Default is 9.8485 * 10^-10}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the initial 208Pb/204Pb for a sample with a known age (t),
measured 206Pb/204Pb (x), measured 207Pb/204Pb (y) and measured 208Pb/204Pb ratios.
}
\examples{
# Example of basic method using a hypothetical sample age of 2700 Ma, measured
# 206Pb/204Pb = 13.5, measured 207Pb/204Pb = 14.5 and measured 208Pb/204Pb = 33
  Calc84in(2700, 13.5, 14.5, 33)

# Example of using additional optional arguments based on Stacey & Kramers (1975) 1st stage model
  Calc84in(2700, 13.5, 14.5, 33, T1 = 4570, X1 = 9.307, Y1 = 10.294, Z1 = 31.23)

}
