% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcKafunction.R
\name{CalcKa}
\alias{CalcKa}
\title{Calculate the time-integrated kappa}
\usage{
# Basic usage accepting default model parameters for Stacey & Kramers (1975) 2nd stage model
  CalcKa(t, z, x, ...)

# Advanced usage with optional arguments
  CalcKa(t, z, x, T1, X1, Z1, L2, L8)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{z}{sample 208Pb/204Pb ratio}

\item{x}{sample 206Pb/204Pb ratio}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Z1}{208Pb/204Pb model starting composition. Default is 31.23}

\item{L2}{232Th decay constant. Default is 0.49475 * 10^-10}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the time-integrated kappa
}
\examples{
# Example of basic method using a hypothetical sample age of 2700 Ma, measured
# 208Pb/204Pb = 33 and measured 206Pb/204Pb = 13.5
  CalcKa(2700, 33, 13.5)

# Example of using optional arguments based on Stacey & Kramers (1975) 1st stage model
  CalcKa(2700, 33, 13.5, T1 = 4570, X1 = 9.307, Z1 = 31.23)

}
