% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochron76slopefunction.R
\name{geochron76slope}
\alias{geochron76slope}
\title{Slope of a geochron line on a 206Pb/204Pb vs 207Pb/204Pb plot for time (t)}
\usage{
# Calculate the geochron on a 206Pb/204Pb vs 207Pb/204Pb plot for time = 2700 Ma
  geochron76slope(2700)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Y1}{207Pb/204Pb model starting composition. Default is 12.998}

\item{Mu1}{238U/204Pb model. Default is 9.74}

\item{U8U5}{Present-day 238U/235U ratio. Default is 137.88}

\item{L5}{235U decay constant. Default is 9.8485 * 10^-10}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the geochron slope on a 206Pb/204Pb vs 207Pb/204Pb plot for a given time (t)
}
\examples{
# Calculate the geochron on a 206Pb/204Pb vs 207Pb/204Pb plot for time = 2700 Ma
  geochron76slope(2700)
}
