% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochron86slopefunction.R
\name{geochron86slope}
\alias{geochron86slope}
\title{Slope of a geochron line on a 206Pb/204Pb vs 208Pb/204Pb plot for time (t)}
\usage{
# Calculate the geochron on a 206Pb/204Pb vs 207Pb/204Pb plot for time = 2700 Ma
  geochron86slope(2700)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Z1}{208Pb/204Pb model starting composition. Default is 31.23}

\item{Mu1}{238U/204Pb model. Default is 9.74}

\item{W1}{232Th/204Pb model. Default is 36.84}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}

\item{L2}{232Th decay constant. Default is 0.49475 * 10^-10}
}
\description{
This function calculates the geochron slope on a 206Pb/204Pb vs 208Pb/204Pb plot for a given time (t)
}
\examples{
# Calculate the geochron on a 206Pb/204Pb vs 207Pb/204Pb plot for time = 2700 Ma
  geochron86slope(2700)
}
