% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastRadiogenicfunction.R
\name{LeastRad}
\alias{LeastRad}
\title{Calculate the least radiogenic values}
\usage{
LeastRad(df, group, value)
}
\arguments{
\item{df}{original dataframe from which the data is to be filtered}

\item{group}{group e.g. deposit, region}

\item{value}{e.g. 206Pb/204Pb}
}
\description{
This function filters a dataset for the least radiogenic analysis
}
\examples{
# Example
  dfLR <- LeastRad(df, DepositName, Pb74)

}
