% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcurvefunction.R
\name{modelcurve}
\alias{modelcurve}
\title{Create a dataframe with values to produce model curves}
\usage{
# Basic method accepting default model parameters for Stacey & Kramers (1975)
# 2nd stage model
  modelcurve(t, ...)

# Advanced method using other values for optional parameters
  modelcurve(t, T1, X1, Y1, Z1, Mu1, W1, U8U5, L5, L8, L2)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{T1}{model start time (Ma). Default is 3700}

\item{X1}{206Pb/204Pb model starting composition. Default is 11.152}

\item{Y1}{207Pb/204Pb model starting composition. Default is 12.998}

\item{Z1}{208Pb/204Pb model starting composition. Default is 31.23}

\item{Mu1}{238U/204Pb model. Default is 9.74}

\item{W1}{232Th/204Pb model. Default is 36.84}

\item{U8U5}{Present-day 238U/235U ratio. Default is 137.88}

\item{L5}{235U decay constant. Default is 9.8485 * 10^-10}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}

\item{L2}{232Th decay constant. Default is 0.49475 * 10^-10}
}
\description{
This function creates a dataframe with four columns (time, 206Pb/204Pb, 207Pb/204Pb and 208Pb/204Pb ratios). These can be used for plotting model curves.
}
\examples{
# Create a dataframe 'df' with the default Stacey & Kramers (1975) 2nd stage values
# for the time period 3700-0 Ma; which can then be used to plot a model curve
  df <- modelcurve(0:3700)

# Create a dataframe 'df' with the Stacey & Kramers (1975) 1st stage model values
# over the time period 3700-0 Ma.
  df <- modelcurve(0:3700, T1 = 4570, X1 = 9.307, Mu1 = 7.19)

# Create a dataframe 'df' for a model source Mu value of 8, over the time period
# 3700-0 Ma.
  df <- modelcurve(0:3700, Mu1 = 8)
}
