% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mslopefunction.R
\name{mslope}
\alias{mslope}
\title{Calculate the slope for a given sample age (t)}
\usage{
mslope(t, ..., U8U5 = 137.88, L5 = 9.8485 * 10^-10, L8 = 1.55125 * 10^-10)
}
\arguments{
\item{t}{sample age or time (Ma)}

\item{...}{optional arguments}

\item{U8U5}{Present-day 238U/235U ratio. Default is 137.88}

\item{L5}{235U decay constant. Default is 9.8485 * 10^-10}

\item{L8}{238U decay constant. Default is 1.55125 * 10^-10}
}
\description{
This function calculates the slope of a line from which samples from the same reservoir, at the same time, will fall along
}
\examples{
# Example of basic method using a hypothetical sample age of 2700 Ma
  mslope(2700)
}
